requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
kafka-python==1.4.2
redis==3.0.1
requests==2.18.4
schedule==0.6.0
urllib3==1.22
pytz==2018.9
python-dateutil==2.8.0
xmltodict==0.12.0